Fixing an error when running step 2-3 with count variables
After running step 2-3 (after the training, to generate the synthetic records) with count variables, all the values of my synthetic records were between 0 and 1, thus binary after rounding. It seems that the default value of `medgan.py` is `--data_type="binary"`  so we need to add the argument `--data_type="binary"` in the final step.